+++ CHANGES.md
@@ -12,6 +12,8 @@
 * Fixed crash when removing USD prims in certain order.
 * Fixed issue where Cesium ion session would not resume on reload.
 * The movie capture tool now waits for tilesets to complete loading before it captures a frame.
+* Added support for WMS raster overlays.
+* Added raster overlay options: `maximumScreenSpaceError`, `maximumTextureSize`, `maximumSimultaneousTileLoads`, `subTileCacheBytes`.
 
 ### v0.17.0 - 2024-02-01